Create appointment by nutritionist

+++ src/services/nutrition/deleteMealPlan.service.ts
@@ -1,9 +1,9 @@
 
-import { mealPlanRepository } from "../../repositories/mealPlanRepository.js";
-import { verifyNutritionistAccess } from "./nutritionAccess.service.js";
+import { mealPlanRepository } from "../../repositories/mealPlanRepository";
+import { verifyNutritionistAccess } from "./nutritionAccess.service";
 
 export const deleteMealPlanService = async(userId: string, mealPlanId: string) => {
-    const user = await verifyNutritionistAccess(userId);
+     await verifyNutritionistAccess(userId);
     const mealPlan = await mealPlanRepository.findById(mealPlanId);
     if(!mealPlan) throw new Error("Meal plan not found");
     return await mealPlanRepository.delete(mealPlanId)
